refactor: combine extractNotes
