+++ lib/v2/qoo-app/utils.js
@@ -1,3 +1,6 @@
+const { parseDate } = require('@/utils/parse-date');
+const timezone = require('@/utils/timezone');
+
 const appsUrl = 'https://apps.qoo-app.com';
 const newsUrl = 'https://news.qoo-app.com';
 const notesUrl = 'https://notes.qoo-app.com';
@@ -23,6 +26,25 @@ const fixImg = ($) => {
     });
 };
 
+const extractNotes = ($) => {
+    $('.qoo-note-wrap')
+        .toArray()
+        .map((item) => {
+            item = $(item);
+            const title = item.find('.qoo-note-view .content-title').text();
+            const author = item.find('cite.name').text();
+            const pubDate = timezone(parseDate(item.find('time').text(), 'YYYY-MM-DD HH:mm'), 8);
+            item.find('cite.name, time, footer').remove();
+            return {
+                title,
+                description: item.find('.qoo-note-view').html(),
+                link: item.find('a.link-wrap').attr('href'),
+                pubDate,
+                author,
+            };
+        });
+};
+
 module.exports = {
     appsUrl,
     newsUrl,
@@ -31,4 +53,5 @@ module.exports = {
     userUrl,
     siteIcon,
     fixImg,
+    extractNotes,
 };
